refactor: split persistence utilities
