+++ tests/duke/persist_golden_test.cpp
@@ -1,4 +1,4 @@
-#include "core/persist.h"
+#include "core/MaterialDTO.h"
 #include <cassert>
 #include <fstream>
 #include <sstream>
