+++ tests/duke/persist_io_test.cpp
@@ -1,4 +1,4 @@
-#include "core/persist.h"
+#include "core/MaterialDTO.h"
 #include <vector>
 #include <cassert>
 #include <filesystem>
